VHAR-8178 Poista vanhat Jenkins-pipelinen tiedostot
